+++ src/main/scala/clases/InvestigadorBarserker.scala
@@ -3,7 +3,7 @@ package clases
 import mixins.luchador.Barserker
 
 
-class InvestigadorBarserker(override val vidaInicial: Int,override val corduraInicial: Int)
+class InvestigadorBarserker(override val vidaInicial: Double,override val corduraInicial: Int)
   extends Investigador(vidaInicial,corduraInicial) with Barserker{
 
 }
