+++ src/main/scala/clases/InvestigadorCobarde.scala
@@ -3,7 +3,7 @@ package clases
 import mixins.luchador.Cobarde
 
 
-class InvestigadorCobarde(override val vidaInicial: Int,override val corduraInicial: Int)
+class InvestigadorCobarde(override val vidaInicial: Double,override val corduraInicial: Int)
   extends Investigador(vidaInicial,corduraInicial) with Cobarde{
 
 }
